+++ CHANGELOG.md
@@ -15,6 +15,10 @@
 - `DebugConfig.setDebugLoggingEnabled` now also controls JS-side debug logging; `console.log` calls in the `Network` and `Drm` modules are suppressed unless debug logging is enabled
 - Update Bitmovin's native iOS SDK version to `3.111.0`
 
+### Fixed
+
+- Android: Picture in Picture fragmented rendering when resizing the PiP window
+
 ## [1.15.0] - 2026-03-27
 
 ### Added